Fix duplicate network request entry during connection
When connecting to a network request, a new NetworkRequestEntry is
created by WifiPickerTracker alongside the unsaved regular network,
which results in two of the same network appearing during the L2
connection phase. To fix this, filter out any disconnected networks that
have the same SSID and security type as the currently connecting
network.

Bug: 292064115
Test: manually verify no duplicate network appears when connecting to
network request

Change-Id: I3d6abb0d0a1e4eeab1b8fa396660851ba090c611

Former-commit-id: ae485a6774d29dc7efa0ddb52155550825029604