Fix flash notification preview not working for second user issue
When switching to second user and try to preview the flash notification
effect it'll not work. The root cause is the receiver is registered in
system process and the broadcast is sent as second user. Needs to send
the broadcast as system user to let the receiver receive the preview
broadcast.

Bug: 286039180
Test: atest FlashNotificationsPreviewPreferenceControllerTest
Test: switch to second user and test the preview function
Change-Id: I7dd553dac9fd66dab59eaec25849aac9f92989b6

Former-commit-id: 65a9173d575b2f5e14a19c95bdee87554a95c53c