Skip test broken by @NonNull annotation fix
b/294110802 fixed the @NonNull annotation so that it was visible to
kotlin when applied to methods outside the current build module.  This
caused kotlin to insert more null checks, which broke a test that was
mocking a method annoated with @NonNull and forcing it to return null.
Add @Ignore to the test for now.

Bug: 294110802
Bug: 297923588
Test: atest SettingsSpaUnitTests
Change-Id: I86a677c5ac01bcfaeebe8c339791cca5b2b97fc7

Former-commit-id: 4700dd261fea9e4a29db7f352911cf3553ae7cd5