Fixed a11y announcement while setting a PIN
This CL removes explicitly setting the disabled state of the confirm
button while reconfirming the PIN in the setting of PIN flow. We do this
since otherwise a11y services announces the state of the button which is
confusing for users.

Bug: b/281953347
Test: Tested by building and installing Settings APK on local device,
and checking a11y announcement. Also setup a PIN to see if the setup
flow is working properly.

Change-Id: I91b592310e240c3994049330c8a881740a3a0d59

Former-commit-id: 0c7ac1bdc6679465e3274d5951b2b89a8c82de4c