Add ResetMobileNetworkSettingsActivity into Settings
The feature introduces an API (an activity action) that
allow applications to launch the Reset mobile network settings
page in Settings Reset options menu.

Since the page is not implemented as an activity and can't directly
work as a intent filter of such activity action, a trampoline activity
is introduced to service as the default one.

Bug: 271921464
Test: atest TelephonyManagerTest
Test: Feature functionality test with flag on and off
Change-Id: Idfba4bedaac8841205424ecdc6300042c2e94733

Former-commit-id: b010448689c883a6a2d7076d86e4612dfb7f6a16