Add the condition to show the eSIM convert menu.
When pSIM converts to eSIM, the new eSIM is downloaded and pSIM is disabled.
However, if the pSIM is in the preferred SIM, a pop-up occurs that says the downloaded eSIM is preferred.
This may cause confusion for users regarding seamless pSIM conversion.
Therefore, change the condition to show the convert menu if there is a pSIM, pSIM conversion is supported, and there is no active eSIM.

Bug: b/314881248
Test: manual done
Change-Id: I10134057a9ae7845ab9fb8b9b8b9f85629d33fa0

Former-commit-id: 036ed3fceecbf1f6fbbe73105f4cc64a9b431e22