Settings: Profiles inherit call restrictions
Currently, if Settings applies CALLS or SMS restrictions to a user, that
restriction doesn't propogate to its child profiles. It should.

This is actually a no-op right now, since it's not possible to have a
profile on a user that isn't the Admin user anyway. But one day such a
thing may be allowed, so this implementation is more future-proof and
safer.

Test: Manual verification that applying the restriction continues to
work properly a full user
Test: UserDetailsSettingsTest
Bug: 261469887

Change-Id: Ib19cecbd37b6efc90f9655565295e7856427f049

Former-commit-id: 84bc8a918d8ae9c261dbe4b24fffab6dfd4f61ae