Auth-bound keys usability: Update strings
Update the strings for the warning shown to the user when they are about
to remove their screen lock and there are authentication-bound keys
that would be invalidated.

These strings are provided by Android UXW.

Additionally, apply the new string to all types of device lock screen:
Pattern, password and unknown.

Bug: 302109605
Test: Manual, flashed a device and added different types of screen lock.
Change-Id: Ida6f5f16c5aa1671f3f2c1358160b8173a1d1407

Former-commit-id: 47973b88acfbb93ae4471bfc108859f26f50803b