Pass the package name from the intent to the picker fragment
When an app uses the legacy autofill intent to launch the
settings activity we should pass the package name of the
launching app to the picker. If the app is then picked by
the user we will return result_ok to the app and if it
is not picked we will return result_cancel.

Change-Id: I99dbfad2c69ed57ff891c189f726ecc408d9d188
Test: ondevice manual
Bug: 324197609

Former-commit-id: 13c6b078680bbfa2bbad1fd48ce9b36cd2357254