fix(accessibility suw): fix unexpected large padding in accessibility suw when force-two-pane mode
Like the screenshots in b/323771329#comment26, the headerLayout padding is too large when
force-two-pane mode. Therefore, as a short-term fix, we make the AccessibilitySetupWizardUtils
not to adjust the headerLayout padding when shouldForceTwoPane, which
described in b/323771329#comment30. The long-term goal is to remove the
padding adjustment in AccessibilitySetupWizardUtils after the the
padding is handled in SUW library internally.

Bug: 323771329
Flag: NONE
Test: manually
Change-Id: Ie4585d9ee352ca74733d4bc14e957bf901347fdc

Former-commit-id: f85bca368c255591ad4038caf0bc177f5ad76055